Homework7  I need help
